requirements.txt
alembic==1.4.2
azure-core==1.5.0
azure-storage-blob==12.3.1
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
click==6.7
cryptography==2.9.2
dnspython==1.16.0
email-validator==1.1.0
Flask==1.0.2
Flask-Login==0.5.0
Flask-Migrate==2.5.3
Flask-SQLAlchemy==2.4.1
Flask-Uploads==0.2.1
Flask-WTF==0.14.3
idna==2.9
isodate==0.6.0
itsdangerous==0.24
Jinja2==2.10
Mako==1.1.2
MarkupSafe==1.0
msrest==0.6.14
oauthlib==3.1.0
pycparser==2.20
pyodbc==4.0.30
python-dateutil==2.8.1
python-editor==1.0.4
requests==2.23.0
requests-oauthlib==1.3.0
six==1.14.0
SQLAlchemy==1.3.17
urllib3==1.25.9
Werkzeug==0.15.5
WTForms==2.3.1
